requirements.txt
asgiref==3.5.2
Babel==2.11.0
bleach==5.0.1
distlib==0.3.6
dj-static==0.0.6
Django==4.0.5
django-crispy-forms==1.14.0
django-gst-field==0.0.3
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
django-summernote==0.8.20.0
filelock==3.8.0
phonenumbers==8.12.57
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.6
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
virtualenv==20.16.5
webencodings==0.5.1
